Extract update notices into src/core/update-notice.ts
Move release metadata reading and moderator update-notice build/dismiss
into a dedicated module. Typecheck, 167 tests, and build all pass.
